requirements.txt
asgiref==3.8.1
azure-core==1.30.2
azure-storage-blob==12.20.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
Django==5.0.6
djangorestframework==3.15.1
drf-yasg==1.21.7
idna==3.7
inflection==0.5.1
isodate==0.6.1
packaging==24.0
pillow==10.3.0
pycparser==2.22
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
APScheduler==3.10.4